fix static_cast, better buttons positions

+++ algorithms/alt.hpp
@@ -91,7 +91,7 @@ public:
     std::vector<VisualizationEvent> events(landmarks.size() +
                                            result.visualization_events.size());
     for (int i = 0; i < landmarks.size(); ++i) {
-      events[i] = {VisualizationEventType::LANDMARK, landmarks[i]};
+      events[i] = {VisualizationEventType::LANDMARK, static_cast<uint64_t>(landmarks[i])};
     }
     for (int i = 0; i < result.visualization_events.size(); ++i) {
       events[landmarks.size() + i] = result.visualization_events[i];
